Extracted context code to its own IContext file used to set global context variables through custom hooks
