+++ app/layout/index.js
@@ -2,3 +2,4 @@ import Layout from './Layout';
 
 export default Layout;
 export { default as Container } from './shared/Container';
+export { default as useIcon } from './shared/IContext';
